+++ R/data.R
@@ -10,7 +10,7 @@
 #' function asks to download the files to the user's cache directory so that
 #' they can be accessed more quickly in the future.
 #'
-#' @format ## `get_solution("PT-ZB18a")`
+#' @format ## `get_solution("full-ZB18a")`
 #' A data frame with 250,001 rows and 20 columns:
 #' \describe{
 #'   \item{t}{Time \eqn{t} (days).}
@@ -56,9 +56,10 @@
 #' }
 #' @source
 #'
-#' * All astronomical solutions by Zeebe can be found on
+#' All astronomical solutions by Zeebe can be found on
 #'   <http://www.soest.hawaii.edu/oceanography/faculty/zeebe_files/Astro.html>.
-#'   They can be loaded into R quickly, using [get_solution()].
+#'
+#' They can be loaded into R quickly, using [get_solution()].
 #'
 #' @seealso [prepare_solution()]
 #'
@@ -71,8 +72,8 @@
 #'   paleo-applications utilizing obliquity and precession cycles: The role of
 #'   dynamical ellipticity and tidal dissipation. _Paleoceanography and
 #'   Paleoclimatology_. \doi{10.1029/2021PA004349}
-#' @name PT_ZB18a
-#' @aliases PT-ZB18a
+#' @name full_ZB18a
+#' @aliases full-ZB18a
 NULL
 
 #' Astronomical Solutions ZB17 for the past 100 Myr
@@ -84,7 +85,7 @@ NULL
 #'   \item{ecc}{Eccentricity \eqn{e} (unitless).}
 #'   \item{inc}{Inclination \eqn{I} (degrees).}
 #' }
-#' @inherit PT_ZB18a source
+#' @inherit full_ZB18a source
 #' @references
 #' Zeebe, R. E. (2017). Numerical Solutions for the orbital motion of the Solar
 #'   System over the Past 100 Myr: Limits and new results. _The Astronomical
@@ -102,12 +103,13 @@ NULL
 #'   \item{ecc}{Eccentricity \eqn{e} (unitless).}
 #'   \item{inc}{Inclination \eqn{I} (degrees).}
 #' }
-#' @inherit PT_ZB18a source
+#' @inherit full_ZB18a source
 #' @references
 #' Zeebe, R. E., & Lourens, L. J. (2019). Solar System chaos and the
 #'   Paleocene–Eocene boundary age constrained by geology and astronomy.
 #'   _Science_, 365(6456), 926–929. \doi{10.1126/science.aax0612}.'
-#' @name ZB18a-100
+#' @name ZB18a_100
+#' @aliases "ZB18a-100"
 NULL
 
 #' Astronomical Solution ZB18a for the Past 300 Myr
@@ -119,16 +121,17 @@ NULL
 #'   \item{ecc}{Eccentricity \eqn{e} (unitless).}
 #'   \item{inc}{Inclination \eqn{I} (degrees).}
 #' }
-#' @inherit PT_ZB18a source
+#' @inherit full_ZB18a source
 #' @references
 #' Zeebe, R. E., & Lourens, L. J. (2019). Solar System chaos and the
 #'   Paleocene–Eocene boundary age constrained by geology and astronomy.
 #'   _Science_, 365(6456), 926–929. \doi{10.1126/science.aax0612}.'
+#'
 #' Zeebe, R. E. and Lourens, L. J. (2022). Geologically constrained
 #'   astronomical solutions for the Cenozoic era. _Earth and Planetary Science
 #'   Letters_. \doi{10.1016/j.epsl.2022.117595}
-#' @name ZB18a-300
-#' @aliases ZB18a
+#' @name ZB18a_300
+#' @aliases ZB18a "ZB18a-300"
 NULL
 
 #' Astronomical Solutions ZB20 for the past 300 Myr
@@ -140,7 +143,7 @@ NULL
 #'   \item{ee}{Eccentricity \eqn{e} (unitless).}
 #'   \item{inc}{Inclination \eqn{I} (degrees).}
 #' }
-#' @inherit PT_ZB18a source
+#' @inherit full_ZB18a source
 #' @references
 #' Zeebe, R. E. and Lourens, L. J. (2022). Geologically constrained
 #'   astronomical solutions for the Cenozoic era. _Earth and Planetary Science